requirements.txt
pyserial>=3.5
tqdm>=4.65.0
paho-mqtt>=1.6.1
python-dotenv>=1.0.0
websocket-client>=1.6.1
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
jinja2>=3.1.2
aiofiles>=23.1.0
python-multipart>=0.0.6
websockets>=11.0.3  # Required for FastAPI WebSocket support
requests>=2.31.0
Pillow>=9.0.0